+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 - Added support for configuring individual TTL per DNS record. Thanks to @jaapmarcus!
 - Added support for Ubuntu Server 20.04 LTS.
 - Added the ability to set the php cli version per user (using alias).
+- Added support for resolving ip addresses based on geoip database for Awstats 
 
 ### Bugfixes
 - Disable Apache2 Server Status Module by default.